requirements.txt
amqp==5.2.0
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.1.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.8
defusedxml==0.7.1
Django==4.2.13
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==24.2
django-ninja==1.1.0
django-ninja-extra==0.20.7
django-ninja-jwt==5.3.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-simpleui==2024.4.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.1.1
eventlet==0.36.1
fastjsonschema==2.18.1
fqdn==1.5.1
greenlet==3.0.3
idna==3.7
injector==0.21.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itypes==1.2.0
Jinja2==3.1.4
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kombu==5.3.7
MarkupSafe==2.1.5
mistune==3.0.2
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
ninja==1.11.1.1
ninja-schema==0.13.6
notebook==7.0.6
notebook_shim==0.2.3
overrides==7.4.0
pandocfilters==1.5.0
pillow==10.3.0
pillow-avif-plugin==1.4.3
prometheus-client==0.18.0
prompt_toolkit==3.0.47
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
PyMySQL==1.1.1
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pywinpty==2.0.12
PyYAML==6.0.1
qtconsole==5.4.4
QtPy==2.4.1
redis==5.0.6
referencing==0.30.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
